Moved has_course_access from CMS to student.auth (common)
